+++ tests/integ/snowflake/ml/test_utils/spcs_integ_test_base.py
@@ -2,17 +2,11 @@ import uuid
 
 from absl.testing import absltest
 
-from snowflake.ml._internal.utils import snowflake_env
 from snowflake.ml.utils import connection_params
 from snowflake.snowpark import Session
-from tests.integ.snowflake.ml.test_utils import db_manager, test_env_utils
+from tests.integ.snowflake.ml.test_utils import db_manager
 
 
-@absltest.skipUnless(
-    test_env_utils.get_current_snowflake_cloud_type()
-    in [snowflake_env.SnowflakeCloudType.AWS, snowflake_env.SnowflakeCloudType.AZURE],
-    "SPCS only available in AWS and Azure",
-)
 class SpcsIntegTestBase(absltest.TestCase):
     _TEST_CPU_COMPUTE_POOL = "REGTEST_INFERENCE_CPU_POOL"
     _TEST_GPU_COMPUTE_POOL = "REGTEST_INFERENCE_GPU_POOL"